KZ7.2 - 3

+++ quadratic/views.py
@@ -3,7 +3,7 @@ import math
 from quadratic.forms import QuadraticForm
 
 def get_value(data, name):
-    return int(data.get(name))
+    return float(data.get(name))
 
 def discriminant(data):
